Now correctly utilizing incoming colors.  To accomplish this, had to adjust the placement of RGB in the color map to BGR to match what was required of the 8_8_8_8_REV voxel component format.  Better (future) to change this ordering back to RGB and find some way to indicate to the TextureMediator that it should use a different (non REV) voxel component format.
Former-commit-id: dbf9b16b6c47740739c080fb51fd35bc90847dce

+++ console/src/main/java/org/janelia/it/FlyWorkstation/gui/viewer3d/masking/RenderMappingI.java
@@ -14,11 +14,11 @@ import java.util.Map;
  * Supplies a color mapping, given a list of
  */
 public interface RenderMappingI {
-    public static final int FRAGMENT_RENDERING = 1;
-    public static final int NON_RENDERING = 0;
-    public static final int COMPARTMENT_RENDERING = 2;
-    public static final int SOLID_COMPARTMENT_RENDERING = 3;
-    public static final int PASS_THROUGH_RENDERING = 4;
+    public static final byte FRAGMENT_RENDERING = 1;
+    public static final byte NON_RENDERING = 0;
+    public static final byte COMPARTMENT_RENDERING = 2;
+    public static final byte SOLID_COMPARTMENT_RENDERING = 3;
+    public static final byte PASS_THROUGH_RENDERING = 4;
 
     Map<Integer,byte[]> getMapping( Collection<RenderableBean> fragments );
 }
